implement #1425
