+++ src/main/java/gregtech/common/metatileentities/electric/multiblockpart/MetaTileEntityRotorHolder.java
@@ -15,6 +15,7 @@ import gregtech.api.render.Textures;
 import gregtech.api.unification.material.type.IngotMaterial;
 import gregtech.common.items.behaviors.TurbineRotorBehavior;
 import gregtech.common.metatileentities.multi.electric.generator.MetaTileEntityLargeTurbine;
+import gregtech.common.metatileentities.multi.electric.generator.TurbineMultiblockController;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
@@ -76,15 +77,19 @@ public class MetaTileEntityRotorHolder extends MetaTileEntityMultiblockPart impl
             this.frontFaceFree = checkTurbineFaceFree();
         }
 
-        MetaTileEntityLargeTurbine controller = (MetaTileEntityLargeTurbine) getController();
+        TurbineMultiblockController controller = (TurbineMultiblockController) getController();
         boolean isControllerActive = controller != null && controller.isActive();
 
         if (!isHasRotor()) {
             resetRotorSpeed();
         } else if (currentRotorSpeed < maxRotorSpeed && isControllerActive) {
-            incrementSpeed(1);
+            incrementSpeed(controller.getRotorIncrementSpeed());
         } else if (currentRotorSpeed > 0 && !isControllerActive) {
-            incrementSpeed(-3);
+            if(controller == null){
+                incrementSpeed( maxRotorSpeed/10);
+            }else {
+                incrementSpeed(controller.getRotorDecrementSpeed());
+            }
         }
     }
